Bot_Telegram/requirements.txt
python-telegram-bot
telegram
numpy
python-dotenv
sympy
matplotlib